+++ models/problem_model/CodeSubmission.go
@@ -3,5 +3,5 @@ package problem_model
 type CodeSubmission struct {
 	Language  string `json:"language"`
 	Code      string `json:"code"`
-	ProblemID int64  `json:"problem_id"`
+	ProblemID uint   `json:"problem_id"`
 }
